Implement notes for datapoints

+++ src/main/java/com/balpos/app/web/rest/NoteResource.java
@@ -1,42 +1,41 @@
 package com.balpos.app.web.rest;
 
-import com.codahale.metrics.annotation.Timed;
-import com.balpos.app.domain.Note;
+import com.balpos.app.domain.User;
 import com.balpos.app.service.NoteService;
+import com.balpos.app.service.dto.NoteDTO;
+import com.balpos.app.service.util.UserResourceUtil;
 import com.balpos.app.web.rest.util.HeaderUtil;
-import com.balpos.app.web.rest.util.PaginationUtil;
-import io.swagger.annotations.ApiParam;
+import com.codahale.metrics.annotation.Timed;
 import io.github.jhipster.web.util.ResponseUtil;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.net.URISyntaxException;
-
+import java.security.Principal;
+import java.time.LocalDate;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * REST controller for managing Note.
  */
 @RestController
+@Slf4j
 @RequestMapping("/api")
 public class NoteResource {
-
-    private final Logger log = LoggerFactory.getLogger(NoteResource.class);
 
     private static final String ENTITY_NAME = "note";
 
     private final NoteService noteService;
+    private final UserResourceUtil userResourceUtil;
 
-    public NoteResource(NoteService noteService) {
+    public NoteResource(NoteService noteService, UserResourceUtil userResourceUtil) {
         this.noteService = noteService;
+        this.userResourceUtil = userResourceUtil;
     }
 
     /**
@@ -48,66 +47,35 @@ public class NoteResource {
      */
     @PostMapping("/notes")
     @Timed
-    public ResponseEntity<Note> createNote(@RequestBody Note note) throws URISyntaxException {
+    public ResponseEntity<NoteDTO> createNote(@RequestBody NoteDTO note, Principal principal) throws URISyntaxException {
         log.debug("REST request to save Note : {}", note);
+        User user = userResourceUtil.getUserFromLogin(principal.getName()).get();
         if (note.getId() != null) {
-            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "idexists", "A new note cannot already have an ID")).body(null);
+            NoteDTO noteFromDb = noteService.findOne(note.getId(), user);
+            if (noteFromDb == null) {
+                return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "idexists", "A new note cannot already have an ID")).body(null);
+            }
+            note = noteFromDb;
         }
-        Note result = noteService.save(note);
+        NoteDTO result = noteService.save(note, user);
         return ResponseEntity.created(new URI("/api/notes/" + result.getId()))
             .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
             .body(result);
     }
 
     /**
-     * PUT  /notes : Updates an existing note.
-     *
-     * @param note the note to update
-     * @return the ResponseEntity with status 200 (OK) and with body the updated note,
-     * or with status 400 (Bad Request) if the note is not valid,
-     * or with status 500 (Internal Server Error) if the note couldn't be updated
-     * @throws URISyntaxException if the Location URI syntax is incorrect
-     */
-    @PutMapping("/notes")
-    @Timed
-    public ResponseEntity<Note> updateNote(@RequestBody Note note) throws URISyntaxException {
-        log.debug("REST request to update Note : {}", note);
-        if (note.getId() == null) {
-            return createNote(note);
-        }
-        Note result = noteService.save(note);
-        return ResponseEntity.ok()
-            .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, note.getId().toString()))
-            .body(result);
-    }
-
-    /**
      * GET  /notes : get all the notes.
-     *
-     * @param pageable the pagination information
-     * @return the ResponseEntity with status 200 (OK) and the list of notes in body
      */
     @GetMapping("/notes")
     @Timed
-    public ResponseEntity<List<Note>> getAllNotes(@ApiParam Pageable pageable) {
+    public ResponseEntity<Map<String, List<NoteDTO>>> getAllNotes(@RequestParam LocalDate dateOfNote, Principal principal) {
         log.debug("REST request to get a page of Notes");
-        Page<Note> page = noteService.findAll(pageable);
-        HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/notes");
-        return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
-    }
+        User user = userResourceUtil.getUserFromLogin(principal.getName()).get();
+        List<NoteDTO> notes = noteService.findAll(user, dateOfNote);
 
-    /**
-     * GET  /notes/:id : get the "id" note.
-     *
-     * @param id the id of the note to retrieve
-     * @return the ResponseEntity with status 200 (OK) and with body the note, or with status 404 (Not Found)
-     */
-    @GetMapping("/notes/{id}")
-    @Timed
-    public ResponseEntity<Note> getNote(@PathVariable Long id) {
-        log.debug("REST request to get Note : {}", id);
-        Note note = noteService.findOne(id);
-        return ResponseUtil.wrapOrNotFound(Optional.ofNullable(note));
+        Map<String, List<NoteDTO>> groupedNotes = notes.stream().collect(Collectors.groupingBy(NoteDTO::getCategory));
+
+        return ResponseUtil.wrapOrNotFound(Optional.ofNullable(groupedNotes));
     }
 
     /**
@@ -118,8 +86,13 @@ public class NoteResource {
      */
     @DeleteMapping("/notes/{id}")
     @Timed
-    public ResponseEntity<Void> deleteNote(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteNote(@PathVariable Long id, Principal principal) {
         log.debug("REST request to delete Note : {}", id);
+        User user = userResourceUtil.getUserFromLogin(principal.getName()).get();
+        NoteDTO note = noteService.findOne(id, user);
+        if (note == null) {
+            return ResponseEntity.badRequest().build();
+        }
         noteService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
     }
